% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{assess_env_collinearity}
\alias{assess_env_collinearity}
\title{Assess environmental collinearity}
\usage{
assess_env_collinearity(env_list)
}
\arguments{
\item{env_list}{List of environmental variables}
}
\value{
Matrix of correlations and VIF values
}
\description{
Assess environmental collinearity
}
\keyword{internal}
