% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_gea.R
\name{assess_gradient_quality}
\alias{assess_gradient_quality}
\title{Assess environmental gradient quality}
\usage{
assess_gradient_quality(holobiont, env_variable)
}
\arguments{
\item{holobiont}{Holobiont data object}

\item{env_variable}{Environmental variable name}
}
\value{
List with gradient quality metrics
}
\description{
Assess environmental gradient quality
}
\keyword{internal}
