% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{assess_gradient_strength}
\alias{assess_gradient_strength}
\title{Assess strength of environmental gradient}
\usage{
assess_gradient_strength(env_var)
}
\arguments{
\item{env_var}{Environmental variable vector}
}
\value{
List with gradient assessment
}
\description{
Assess strength of environmental gradient
}
\keyword{internal}
