% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_gea.R
\name{assess_overall_suitability}
\alias{assess_overall_suitability}
\title{Assess overall suitability for GEA}
\usage{
assess_overall_suitability(assumption_results)
}
\arguments{
\item{assumption_results}{List of assumption check results}
}
\value{
Overall assessment
}
\description{
Assess overall suitability for GEA
}
\keyword{internal}
