% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_gea.R
\name{assess_structure_confounding}
\alias{assess_structure_confounding}
\title{Assess population structure confounding}
\usage{
assess_structure_confounding(holobiont, env_variable)
}
\arguments{
\item{holobiont}{Holobiont data object}

\item{env_variable}{Environmental variable name}
}
\value{
List with confounding assessment
}
\description{
Assess population structure confounding
}
\keyword{internal}
