% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{calculate_autocorrelation}
\alias{calculate_autocorrelation}
\title{Calculate autocorrelation in environmental data}
\usage{
calculate_autocorrelation(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Lag-1 autocorrelation
}
\description{
Calculate autocorrelation in environmental data
}
\keyword{internal}
