% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{calculate_data_completeness}
\alias{calculate_data_completeness}
\title{Calculate data completeness}
\usage{
calculate_data_completeness(holobiont)
}
\arguments{
\item{holobiont}{Holobiont data object}
}
\value{
Data completeness score (0-1)
}
\description{
Calculate data completeness
}
\keyword{internal}
