% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_response.R
\name{calculate_environmental_effect}
\alias{calculate_environmental_effect}
\title{Calculate environmental effect}
\usage{
calculate_environmental_effect(env, params)
}
\arguments{
\item{env}{Environmental variable}

\item{params}{Simulation parameters}
}
\value{
Numeric vector of environmental effects
}
\description{
Calculate environmental effect
}
\keyword{internal}
