% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{calculate_evenness}
\alias{calculate_evenness}
\title{Calculate community evenness}
\usage{
calculate_evenness(community_matrix)
}
\arguments{
\item{community_matrix}{Community matrix}
}
\value{
Vector of evenness values (Pielou's J)
}
\description{
Calculate community evenness
}
\keyword{internal}
