% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{calculate_fst}
\alias{calculate_fst}
\title{Calculate Fst between populations}
\usage{
calculate_fst(snps, populations)
}
\arguments{
\item{snps}{SNP matrix}

\item{populations}{Population assignments}
}
\value{
Weighted average Fst
}
\description{
Calculate Fst between populations
}
\keyword{internal}
