% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_response.R
\name{calculate_genetic_value}
\alias{calculate_genetic_value}
\title{Calculate genetic value from genotypes}
\usage{
calculate_genetic_value(snps, causal_loci, effects, params)
}
\arguments{
\item{snps}{SNP matrix}

\item{causal_loci}{Integer vector of causal locus positions}

\item{effects}{Numeric vector of effect sizes}

\item{params}{Simulation parameters}
}
\value{
Numeric vector of genetic values
}
\description{
Calculate genetic value from genotypes
}
\keyword{internal}
