% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{calculate_r2}
\alias{calculate_r2}
\title{Calculate r^2 between two SNPs}
\usage{
calculate_r2(snp1, snp2)
}
\arguments{
\item{snp1}{First SNP vector}

\item{snp2}{Second SNP vector}
}
\value{
r^2 value
}
\description{
Calculate r^2 between two SNPs
}
\keyword{internal}
