% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_required_n}
\alias{calculate_required_n}
\title{Calculate required sample size for target power}
\usage{
calculate_required_n(power_results, target_power = 0.8)
}
\arguments{
\item{power_results}{Object of class 'gea_power_analysis'}

\item{target_power}{Target power threshold (default: 0.8)}
}
\value{
Integer sample size needed, or NA if target not achieved
}
\description{
Calculate required sample size for target power
}
