% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_gea.R
\name{check_lfmm_assumptions}
\alias{check_lfmm_assumptions}
\title{Check LFMM-specific assumptions}
\usage{
check_lfmm_assumptions(holobiont, env_variable)
}
\arguments{
\item{holobiont}{Holobiont data object}

\item{env_variable}{Environmental variable name}
}
\value{
List with LFMM-specific checks
}
\description{
Check LFMM-specific assumptions
}
\keyword{internal}
