% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_gea.R
\name{check_rda_assumptions}
\alias{check_rda_assumptions}
\title{Check RDA-specific assumptions}
\usage{
check_rda_assumptions(holobiont, env_variable)
}
\arguments{
\item{holobiont}{Holobiont data object}

\item{env_variable}{Environmental variable name}
}
\value{
List with RDA-specific checks
}
\description{
Check RDA-specific assumptions
}
\keyword{internal}
