% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_genetics.R
\name{convert_to_snp_matrix}
\alias{convert_to_snp_matrix}
\title{Convert coala simulation output to SNP matrix}
\usage{
convert_to_snp_matrix(coala_sim)
}
\arguments{
\item{coala_sim}{Output from coala::simulate()}
}
\value{
SNP matrix (individuals x loci)
}
\description{
Convert coala simulation output to SNP matrix
}
\keyword{internal}
