% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_gea.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers in holobiont data}
\usage{
detect_outliers(holobiont, env_variable)
}
\arguments{
\item{holobiont}{Holobiont data object}

\item{env_variable}{Environmental variable name}
}
\value{
List with outlier information
}
\description{
Detect outliers in holobiont data
}
\keyword{internal}
