% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_community_matrix}
\alias{diagnose_community_matrix}
\title{Diagnose community matrix}
\usage{
diagnose_community_matrix(community_matrix)
}
\arguments{
\item{community_matrix}{Community matrix (samples x taxa)}
}
\value{
List of community diagnostics
}
\description{
Diagnose community matrix
}
\keyword{internal}
