% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_environmental_data}
\alias{diagnose_environmental_data}
\title{Diagnose environmental data}
\usage{
diagnose_environmental_data(environment)
}
\arguments{
\item{environment}{List of environmental variables}
}
\value{
List of environmental diagnostics
}
\description{
Diagnose environmental data
}
\keyword{internal}
