% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_genetic_data}
\alias{diagnose_genetic_data}
\title{Diagnose genetic data quality}
\usage{
diagnose_genetic_data(snps)
}
\arguments{
\item{snps}{SNP matrix}
}
\value{
List of genetic diagnostics
}
\description{
Diagnose genetic data quality
}
\keyword{internal}
