% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_holobiont_response}
\alias{diagnose_holobiont_response}
\title{Diagnose holobiont response data}
\usage{
diagnose_holobiont_response(holobiont)
}
\arguments{
\item{holobiont}{Holobiont data object}
}
\value{
List of response diagnostics
}
\description{
Diagnose holobiont response data
}
\keyword{internal}
