% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_population_structure}
\alias{diagnose_population_structure}
\title{Diagnose population structure}
\usage{
diagnose_population_structure(holobiont)
}
\arguments{
\item{holobiont}{Holobiont data object}
}
\value{
List of population structure diagnostics
}
\description{
Diagnose population structure
}
\keyword{internal}
