% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{estimate_decay_rate}
\alias{estimate_decay_rate}
\title{Estimate LD decay rate}
\usage{
estimate_decay_rate(distances, r2_values)
}
\arguments{
\item{distances}{Vector of distances between SNP pairs}

\item{r2_values}{Vector of r^2 values}
}
\value{
Decay rate estimate
}
\description{
Estimate LD decay rate
}
\keyword{internal}
