% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{estimate_ld_decay}
\alias{estimate_ld_decay}
\title{Estimate linkage disequilibrium decay}
\usage{
estimate_ld_decay(snps, max_pairs = 10000)
}
\arguments{
\item{snps}{SNP matrix}

\item{max_pairs}{Maximum number of SNP pairs to sample (default: 10000)}
}
\value{
List with LD statistics
}
\description{
Estimate linkage disequilibrium decay
}
\keyword{internal}
