% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_gea.R
\name{estimate_signal_noise}
\alias{estimate_signal_noise}
\title{Estimate signal-to-noise ratio}
\usage{
estimate_signal_noise(holobiont, env_variable)
}
\arguments{
\item{holobiont}{Holobiont data object}

\item{env_variable}{Environmental variable name}
}
\value{
List with SNR estimates
}
\description{
Estimates the ratio of genetic signal to environmental noise in the data.
Higher values indicate better power for detection.
}
\keyword{internal}
