% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{estimate_spatial_autocorrelation}
\alias{estimate_spatial_autocorrelation}
\title{Estimate spatial autocorrelation}
\usage{
estimate_spatial_autocorrelation(env_var)
}
\arguments{
\item{env_var}{Environmental variable}
}
\value{
Spatial autocorrelation estimate
}
\description{
Estimate spatial autocorrelation
}
\keyword{internal}
