% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_gea.R
\name{fast_pca}
\alias{fast_pca}
\title{Fast PCA implementation}
\usage{
fast_pca(snps, n_pcs = 5)
}
\arguments{
\item{snps}{SNP matrix}

\item{n_pcs}{Number of PCs to compute}
}
\value{
List with PCs and variance explained
}
\description{
Fast PCA implementation
}
\keyword{internal}
