% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gea_methods.R, R/power_calculations.R
\name{format_power_results}
\alias{format_power_results}
\title{Format power results for output}
\usage{
format_power_results(power_list, sample_sizes, params)

format_power_results(power_list, sample_sizes, params)
}
\arguments{
\item{power_list}{List of power results for each sample size}

\item{sample_sizes}{Vector of sample sizes tested}

\item{params}{Simulation parameters}
}
\value{
Data frame of class 'gea_power_analysis'

Data frame of class 'gea_power_analysis'
}
\description{
Format power results for output

Format power results for output
}
\keyword{internal}
