% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_response.R
\name{format_response_variable}
\alias{format_response_variable}
\title{Format response variable based on type}
\usage{
format_response_variable(holobiont, response, params)
}
\arguments{
\item{holobiont}{Holobiont data object}

\item{response}{Raw response values}

\item{params}{Simulation parameters}
}
\value{
Updated holobiont object
}
\description{
Format response variable based on type
}
\keyword{internal}
