% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_environment.R
\name{generate_continuous_environment}
\alias{generate_continuous_environment}
\title{Generate continuous environmental gradient}
\usage{
generate_continuous_environment(n, holobiont, params)
}
\arguments{
\item{n}{Number of individuals}

\item{holobiont}{Holobiont data object}

\item{params}{Simulation parameters}
}
\value{
Numeric vector of environmental values
}
\description{
Generate continuous environmental gradient
}
\keyword{internal}
