% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_response.R
\name{generate_effect_sizes}
\alias{generate_effect_sizes}
\title{Generate effect sizes for causal loci}
\usage{
generate_effect_sizes(params)
}
\arguments{
\item{params}{Simulation parameters}
}
\value{
Numeric vector of effect sizes
}
\description{
Generate effect sizes for causal loci
}
\keyword{internal}
