% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_environment.R
\name{generate_multivariate_environment}
\alias{generate_multivariate_environment}
\title{Generate multivariate environmental data}
\usage{
generate_multivariate_environment(n, params)
}
\arguments{
\item{n}{Number of individuals}

\item{params}{Simulation parameters}
}
\value{
Matrix of environmental variables (n x n_variables)
}
\description{
Generate multivariate environmental data
}
\keyword{internal}
