% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holobiont_data.R
\name{get_environment}
\alias{get_environment}
\title{Get environment data from holobiont data}
\usage{
get_environment(holobiont, variable = NULL)
}
\arguments{
\item{holobiont}{A holobiont_data object}

\item{variable}{Character. Specific environmental variable to retrieve (optional)}
}
\value{
List of environmental data or specific variable vector
}
\description{
Get environment data from holobiont data
}
