% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holobiont_data.R
\name{get_host_genotypes}
\alias{get_host_genotypes}
\title{Get host genotypes from holobiont data}
\usage{
get_host_genotypes(holobiont)
}
\arguments{
\item{holobiont}{A holobiont_data object}
}
\value{
SNP matrix
}
\description{
Get host genotypes from holobiont data
}
