% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holobiont_data.R
\name{get_microbiome}
\alias{get_microbiome}
\title{Get microbiome data from holobiont data}
\usage{
get_microbiome(holobiont, type = "fungi")
}
\arguments{
\item{holobiont}{A holobiont_data object}

\item{type}{Character. Type of microbiome data ("fungi", "bacteria", etc.)}
}
\value{
Microbiome data
}
\description{
Get microbiome data from holobiont data
}
