% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot.holobiont_diagnostics}
\alias{plot.holobiont_diagnostics}
\title{Plot holobiont diagnostics}
\usage{
\method{plot}{holobiont_diagnostics}(x, which = "all", ...)
}
\arguments{
\item{x}{A holobiont_diagnostics object}

\item{which}{Character vector. Which plots to generate: "genetic", "environment",
"response", "all" (default: "all")}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
List of ggplot objects (invisible)
}
\description{
Plot holobiont diagnostics
}
