% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_env_individual}
\alias{plot_env_individual}
\title{Plot individual-level environmental distributions}
\usage{
plot_env_individual(ind_env)
}
\description{
Plot individual-level environmental distributions
}
