% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_env_space.R
\name{plot_env_space}
\alias{plot_env_space}
\title{Plot environmental space}
\usage{
plot_env_space(env_space)
}
\arguments{
\item{env_space}{An object created by define_env_space().}
}
\value{
A ggplot object showing realised environmental gradients.
}
\description{
Plot environmental space
}
