% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_environmental_diagnostics}
\alias{plot_environmental_diagnostics}
\title{Plot environmental diagnostics}
\usage{
plot_environmental_diagnostics(env_diag)
}
\arguments{
\item{env_diag}{Environmental diagnostics list}
}
\value{
ggplot object
}
\description{
Plot environmental diagnostics
}
\keyword{internal}
