% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_heterozygosity}
\alias{plot_heterozygosity}
\title{Plot heterozygosity}
\usage{
plot_heterozygosity(genetic_diag)
}
\arguments{
\item{genetic_diag}{Genetic diagnostics list}
}
\value{
ggplot object
}
\description{
Plot heterozygosity
}
\keyword{internal}
