% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_host_model.R
\name{plot_host_model}
\alias{plot_host_model}
\title{Plot host model diagnostics}
\usage{
plot_host_model(host_model)
}
\arguments{
\item{host_model}{Object returned by \code{define_host_model()}.}
}
\value{
A ggplot object with a simple textual summary.
}
\description{
Plot host model diagnostics
}
