% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_maf_distribution}
\alias{plot_maf_distribution}
\title{Plot MAF distribution}
\usage{
plot_maf_distribution(genetic_diag)
}
\arguments{
\item{genetic_diag}{Genetic diagnostics list}
}
\value{
ggplot object
}
\description{
Plot MAF distribution
}
\keyword{internal}
