% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_missing_data}
\alias{plot_missing_data}
\title{Plot missing data}
\usage{
plot_missing_data(genetic_diag)
}
\arguments{
\item{genetic_diag}{Genetic diagnostics list}
}
\value{
ggplot object
}
\description{
Plot missing data
}
\keyword{internal}
