% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_n_detected}
\alias{plot_n_detected}
\title{Plot number of loci detected}
\usage{
plot_n_detected(power_results)
}
\arguments{
\item{power_results}{Object of class 'gea_power_analysis'}
}
\value{
ggplot object
}
\description{
Plot number of loci detected
}
