% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_power_fdr}
\alias{plot_power_fdr}
\title{Plot power and FDR together}
\usage{
plot_power_fdr(power_results)
}
\arguments{
\item{power_results}{Object of class 'gea_power_analysis'}
}
\value{
ggplot object
}
\description{
Plot power and FDR together
}
