% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_power_landscape}
\alias{plot_power_landscape}
\title{Plot power landscape (design exploration)}
\usage{
plot_power_landscape(power_results)
}
\description{
Plot power landscape (design exploration)
}
