% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_response_diagnostics}
\alias{plot_response_diagnostics}
\title{Plot response diagnostics}
\usage{
plot_response_diagnostics(response_diag)
}
\arguments{
\item{response_diag}{Response diagnostics list}
}
\value{
ggplot object or NULL
}
\description{
Plot response diagnostics
}
\keyword{internal}
