% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_gea.R
\name{print.gea_assumptions}
\alias{print.gea_assumptions}
\title{Print method for gea_assumptions}
\usage{
\method{print}{gea_assumptions}(x, ...)
}
\arguments{
\item{x}{A gea_assumptions object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for gea_assumptions
}
