% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holobiont_data.R
\name{print.holobiont_data}
\alias{print.holobiont_data}
\title{Print method for holobiont_data}
\usage{
\method{print}{holobiont_data}(x, ...)
}
\arguments{
\item{x}{A holobiont_data object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for holobiont_data
}
