% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.holobiont_diagnostics}
\alias{print.holobiont_diagnostics}
\title{Print method for holobiont_diagnostics}
\usage{
\method{print}{holobiont_diagnostics}(x, ...)
}
\arguments{
\item{x}{A holobiont_diagnostics object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for holobiont_diagnostics
}
