% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{reshape_cor_matrix}
\alias{reshape_cor_matrix}
\title{Reshape correlation matrix for plotting}
\usage{
reshape_cor_matrix(cor_matrix)
}
\arguments{
\item{cor_matrix}{Correlation matrix}
}
\value{
Data frame for ggplot
}
\description{
Reshape correlation matrix for plotting
}
\keyword{internal}
