% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gea_methods.R
\name{run_gea_analysis}
\alias{run_gea_analysis}
\title{Run GEA analysis}
\usage{
run_gea_analysis(holobiont, params)
}
\arguments{
\item{holobiont}{A holobiont_data object}

\item{params}{Simulation parameters}
}
\value{
List with p-values for each SNP
}
\description{
Run GEA analysis
}
\keyword{internal}
