% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gea_methods.R
\name{run_lfmm}
\alias{run_lfmm}
\title{Run LFMM-based GEA}
\usage{
run_lfmm(snps, env, params)
}
\arguments{
\item{snps}{SNP matrix}

\item{env}{Environmental variable}

\item{params}{Simulation parameters}
}
\value{
Numeric vector of p-values
}
\description{
Run LFMM-based GEA
}
\keyword{internal}
