% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gea_methods.R
\name{run_rda}
\alias{run_rda}
\title{Run RDA-based GEA}
\usage{
run_rda(snps, env, response = NULL)
}
\arguments{
\item{snps}{SNP matrix}

\item{env}{Environmental variable}

\item{response}{Response variable (optional, used for validation)}
}
\value{
Numeric vector of p-values
}
\description{
Run RDA-based GEA
}
\keyword{internal}
