% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gea_methods.R, R/power_calculations.R
\name{run_single_replicate}
\alias{run_single_replicate}
\title{Run single replicate of power simulation}
\usage{
run_single_replicate(n, params)

run_single_replicate(n, params)
}
\arguments{
\item{n}{Sample size for this replicate}

\item{params}{Simulation parameters}
}
\value{
List with power metrics

List with power metrics
}
\description{
Run single replicate of power simulation

Run single replicate of power simulation
}
\keyword{internal}
