% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{sample_snp_pairs}
\alias{sample_snp_pairs}
\title{Sample SNP pairs for LD calculation}
\usage{
sample_snp_pairs(n_snps, n_pairs)
}
\arguments{
\item{n_snps}{Number of SNPs}

\item{n_pairs}{Number of pairs to sample}
}
\value{
Matrix of SNP pair indices
}
\description{
Sample SNP pairs for LD calculation
}
\keyword{internal}
