% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holobiont_data.R
\name{set_environment}
\alias{set_environment}
\title{Set environment data in holobiont data}
\usage{
set_environment(holobiont, variable, values)
}
\arguments{
\item{holobiont}{A holobiont_data object}

\item{variable}{Character. Name of environmental variable}

\item{values}{Numeric vector of values}
}
\value{
Updated holobiont_data object
}
\description{
Set environment data in holobiont data
}
