% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holobiont_data.R
\name{set_host_genotypes}
\alias{set_host_genotypes}
\title{Set host genotypes in holobiont data}
\usage{
set_host_genotypes(holobiont, snp_matrix)
}
\arguments{
\item{holobiont}{A holobiont_data object}

\item{snp_matrix}{Matrix of SNP genotypes}
}
\value{
Updated holobiont_data object
}
\description{
Set host genotypes in holobiont data
}
