% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_response.R
\name{simulate_community_matrix}
\alias{simulate_community_matrix}
\title{Simulate full community matrix}
\usage{
simulate_community_matrix(n, n_taxa, pc1_values, params)
}
\arguments{
\item{n}{Number of individuals}

\item{n_taxa}{Number of taxa}

\item{pc1_values}{PC1 values driving community composition}

\item{params}{Simulation parameters}
}
\value{
Community matrix (individuals x taxa)
}
\description{
Simulate full community matrix
}
\keyword{internal}
