% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_environment.R
\name{simulate_environment_data}
\alias{simulate_environment_data}
\title{Simulate environmental data}
\usage{
simulate_environment_data(holobiont, params)
}
\arguments{
\item{holobiont}{A holobiont_data object}

\item{params}{List of simulation parameters}
}
\value{
Updated holobiont_data object with environment data
}
\description{
Simulate environmental data
}
\keyword{internal}
