% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_genetics.R
\name{simulate_host_genetics}
\alias{simulate_host_genetics}
\title{Simulate host genetics with population structure}
\usage{
simulate_host_genetics(holobiont, params)
}
\arguments{
\item{holobiont}{A holobiont_data object}

\item{params}{List of simulation parameters}
}
\value{
Updated holobiont_data object with host_genotypes
}
\description{
Simulate host genetics with population structure
}
\keyword{internal}
