% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_response.R
\name{simulate_response}
\alias{simulate_response}
\title{Simulate holobiont response variable}
\usage{
simulate_response(holobiont, params)
}
\arguments{
\item{holobiont}{A holobiont_data object}

\item{params}{List of simulation parameters}
}
\value{
Updated holobiont_data object with response data and causal_loci
}
\description{
Simulate holobiont response variable
}
\keyword{internal}
