% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.gea_power_analysis}
\alias{summary.gea_power_analysis}
\title{Summary method for gea_power_analysis}
\usage{
\method{summary}{gea_power_analysis}(object, ...)
}
\arguments{
\item{object}{A gea_power_analysis object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for gea_power_analysis
}
