% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_simulation_parameters}
\alias{validate_simulation_parameters}
\title{Validate simulation parameters}
\usage{
validate_simulation_parameters(...)
}
\arguments{
\item{...}{Named simulation parameters}
}
\value{
List of validated parameters
}
\description{
Validate simulation parameters
}
\keyword{internal}
