% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_variance_decomposition}
\alias{plot_variance_decomposition}
\title{Plot variance decomposition for holobiont components
A unique + important diagnostic!}
\usage{
plot_variance_decomposition(power_results, components)
}
\description{
Plot variance decomposition for holobiont components
A unique + important diagnostic!
}
